feat: Add dynamic data masking
Enhanced the data masking capability to be dynamic, allowing masking by word.

Previously, the microservice only allowed masking parts of the string at the beginning, end, or middle. With this update, it is now possible to mask each word individually according to business requirements.

Example of previous functionality: "Seb*** ****** *****"
Example of current functionality: "Seb**** Cue*** Nav***"

This was achieved by implementing logic that identifies and masks each word in the string according to the new specifications.